+++ guides/Gen 3/Emerald/Method Likelihood Calculator.mdx
@@ -16,7 +16,7 @@
 - Understand why you never encounter your target Pokémon even if you hit the correct advance.
 - Optimize your lead cycle speed to improve your odds of hitting your target Pokémon.
 
-<Wild3MethodDistribution permitEnablingDebugOptions={true} />
+<Wild3MethodDistributionStandalone permitEnablingDebugOptions={true} />
 
 ## Credits
